app/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
huggingface_hub
python-dotenv
requests

